requirements.txt
streamlit>=1.30.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pypdf>=3.17.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
sumy>=0.11.0
nltk>=3.8.1
numpy>=1.26.0
pytest>=7.4.0
fpdf>=1.7.2
plotly>=5.18.0
easyocr>=1.7.1
opencv-python-headless>=4.8.0
